+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+- scrollbar and fixed footer action in sidebar for system theme
+
 ### Removed
 - pinned tab rendering